[fix](arrow) Fix UT DataTypeSerDeArrowTest of Date type #48398

+++ be/src/vec/data_types/serde/data_type_decimal_serde.cpp
@@ -185,9 +185,8 @@ void DataTypeDecimalSerDe<T>::write_column_to_arrow(const IColumn& column, const
 
 template <typename T>
 void DataTypeDecimalSerDe<T>::read_column_from_arrow(IColumn& column,
-                                                     const arrow::Array* arrow_array, int64_t start,
-                                                     int64_t end,
-                                                     const cctz::time_zone& ctz) const {
+                                                     const arrow::Array* arrow_array, int start,
+                                                     int end, const cctz::time_zone& ctz) const {
     auto& column_data = static_cast<ColumnDecimal<T>&>(column).get_data();
     // Decimal<Int128> for decimalv2
     // Decimal<Int128I> for deicmalv3
